LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/font-size-animation-expected.txt


FAIL Animating font-size handled identically for standard and custom properties assert_equals: expected "0px" but got "250px"
